add currency endpoints

+++ src/Domain/Common/Entity.cs
@@ -5,12 +5,12 @@
         /// <summary>
         /// Gets or sets the primary key for this entity.
         /// </summary>
-        public long Id { get; set; }
+        public int Id { get; set; }
     }
     public abstract class Entity : IEntity
     {
         /// <inheritdoc cref="IEntity.Id"/>
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Initializes a new instance of <see cref="Entity"/>.
@@ -21,7 +21,7 @@
         /// Initializes a new instance of <see cref="Entity"/>.
         /// </summary>
         /// <param name="id">The primary key.</param>
-        protected Entity(long id) { Id = id; }
+        protected Entity(int id) { Id = id; }
 
         public override bool Equals(object? obj)
         {
